@@ -1,8 +1,10 @@
 {
   "name": "@octokit/endpoint",
-  "version": "7.0.6",
+  "version": "10.1.1",
+  "type": "module",
   "publishConfig": {
-    "access": "public"
+    "access": "public",
+    "provenance": true
   },
   "description": "Turns REST API endpoints into generic request options",
   "repository": "github:octokit/endpoint.js",
@@ -15,32 +17,36 @@
   "author": "Gregor Martynus (https://github.com/gr2m)",
   "license": "MIT",
   "devDependencies": {
-    "@octokit/tsconfig": "^2.0.0",
+    "@octokit/tsconfig": "^3.0.0",
     "@types/jest": "^29.0.0",
-    "esbuild": "^0.17.19",
+    "@types/node": "^20.10.0",
+    "esbuild": "^0.20.0",
     "glob": "^10.2.7",
     "jest": "^29.0.0",
-    "prettier": "2.8.8",
-    "semantic-release": "^21.0.0",
+    "prettier": "3.2.5",
+    "semantic-release": "^23.0.0",
     "semantic-release-plugin-update-version-in-files": "^1.0.0",
     "ts-jest": "^29.0.0",
     "typescript": "^5.0.0"
   },
   "dependencies": {
-    "@octokit/types": "^9.0.0",
-    "is-plain-object": "^5.0.0",
-    "universal-user-agent": "^6.0.0"
+    "@octokit/types": "^13.0.0",
+    "universal-user-agent": "^7.0.2"
   },
   "engines": {
-    "node": ">= 14"
+    "node": ">= 18"
   },
   "files": [
     "dist-*/**",
     "bin/**"
   ],
-  "main": "dist-node/index.js",
-  "browser": "dist-web/index.js",
-  "types": "dist-types/index.d.ts",
-  "module": "dist-src/index.js",
+  "types": "./dist-types/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist-types/index.d.ts",
+      "import": "./dist-bundle/index.js",
+      "default": "./dist-bundle/index.js"
+    }
+  },
   "sideEffects": false
 }
