@@ -1,49 +1,50 @@
 {
   "name": "@octokit/request-error",
+  "version": "6.1.1",
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
+  "type": "module",
   "description": "Error class for Octokit request errors",
-  "version": "3.0.3",
-  "license": "MIT",
-  "files": [
-    "dist-*/",
-    "bin/"
-  ],
-  "source": "dist-src/index.js",
-  "types": "dist-types/index.d.ts",
-  "main": "dist-node/index.js",
-  "module": "dist-web/index.js",
-  "pika": true,
-  "sideEffects": false,
+  "repository": "github:octokit/request-error.js",
   "keywords": [
     "octokit",
     "github",
     "api",
     "error"
   ],
-  "repository": "github:octokit/request-error.js",
+  "author": "Gregor Martynus (https://github.com/gr2m)",
+  "license": "MIT",
   "dependencies": {
-    "@octokit/types": "^9.0.0",
-    "deprecation": "^2.0.0",
-    "once": "^1.4.0"
+    "@octokit/types": "^13.0.0"
   },
   "devDependencies": {
-    "@pika/pack": "^0.3.7",
-    "@pika/plugin-build-node": "^0.9.0",
-    "@pika/plugin-build-web": "^0.9.0",
-    "@pika/plugin-bundle-web": "^0.9.0",
-    "@pika/plugin-ts-standard-pkg": "^0.9.0",
+    "@octokit/tsconfig": "^3.0.0",
     "@types/jest": "^29.0.0",
-    "@types/node": "^18.0.0",
-    "@types/once": "^1.4.0",
+    "@types/node": "^20.0.0",
+    "esbuild": "^0.20.0",
+    "glob": "^10.2.6",
     "jest": "^29.0.0",
-    "pika-plugin-unpkg-field": "^1.1.0",
-    "prettier": "2.8.3",
+    "prettier": "3.2.5",
     "ts-jest": "^29.0.0",
-    "typescript": "^4.0.0"
+    "typescript": "^5.0.0"
   },
   "engines": {
-    "node": ">= 14"
+    "node": ">= 18"
   },
-  "publishConfig": {
-    "access": "public"
-  }
+  "files": [
+    "dist-*/**",
+    "bin/**"
+  ],
+  "types": "./dist-types/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist-types/index.d.ts",
+      "import": "./dist-src/index.js",
+      "default": "./dist-src/index.js"
+    }
+  },
+  "sideEffects": false,
+  "unpkg": "dist-src/index.js"
 }
