@@ -1,8 +1,10 @@
 {
   "name": "@octokit/request",
-  "version": "6.2.8",
+  "version": "9.1.1",
+  "type": "module",
   "publishConfig": {
-    "access": "public"
+    "access": "public",
+    "provenance": true
   },
   "description": "Send parameterized requests to GitHub's APIs with sensible defaults in browsers and Node",
   "repository": "github:octokit/request.js",
@@ -15,43 +17,44 @@
   "author": "Gregor Martynus (https://github.com/gr2m)",
   "license": "MIT",
   "dependencies": {
-    "@octokit/endpoint": "^7.0.0",
-    "@octokit/request-error": "^3.0.0",
-    "@octokit/types": "^9.0.0",
-    "is-plain-object": "^5.0.0",
-    "node-fetch": "^2.6.7",
-    "universal-user-agent": "^6.0.0"
+    "@octokit/endpoint": "^10.0.0",
+    "@octokit/request-error": "^6.0.1",
+    "@octokit/types": "^13.1.0",
+    "universal-user-agent": "^7.0.2"
   },
   "devDependencies": {
-    "@octokit/auth-app": "^4.0.0",
-    "@octokit/tsconfig": "^2.0.0",
+    "@octokit/auth-app": "^6.0.0",
+    "@octokit/tsconfig": "^3.0.0",
+    "@sinonjs/fake-timers": "^11.2.2",
     "@types/fetch-mock": "^7.2.4",
     "@types/jest": "^29.0.0",
     "@types/lolex": "^5.1.0",
-    "@types/node": "^18.0.0",
-    "@types/node-fetch": "^2.3.3",
+    "@types/node": "^20.0.0",
     "@types/once": "^1.4.0",
-    "esbuild": "^0.18.0",
-    "fetch-mock": "^9.3.1",
+    "@types/sinonjs__fake-timers": "^8.1.5",
+    "esbuild": "^0.20.0",
+    "fetch-mock": "npm:@gr2m/fetch-mock@^9.11.0-pull-request-644.1",
     "glob": "^10.2.4",
     "jest": "^29.0.0",
-    "lolex": "^6.0.0",
-    "prettier": "2.8.8",
+    "prettier": "3.2.5",
     "semantic-release-plugin-update-version-in-files": "^1.0.0",
-    "string-to-arraybuffer": "^1.0.2",
     "ts-jest": "^29.0.0",
     "typescript": "^5.0.0"
   },
   "engines": {
-    "node": ">= 14"
+    "node": ">= 18"
   },
   "files": [
     "dist-*/**",
     "bin/**"
   ],
-  "main": "dist-node/index.js",
-  "browser": "dist-web/index.js",
   "types": "dist-types/index.d.ts",
-  "module": "dist-src/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist-types/index.d.ts",
+      "import": "./dist-bundle/index.js",
+      "default": "./dist-bundle/index.js"
+    }
+  },
   "sideEffects": false
 }
