@@ -60,32 +80,34 @@
     "access": "public"
   },
   "devDependencies": {
-    "@types/mocha": "10.0.0",
+    "@types/mocha": "10.0.6",
     "@types/node": "18.6.5",
-    "@types/sinon": "10.0.13",
-    "@types/webpack": "4.41.26",
+    "@types/sinon": "10.0.20",
+    "@types/webpack": "5.28.5",
     "@types/webpack-env": "1.16.3",
+    "babel-plugin-istanbul": "6.1.1",
     "codecov": "3.8.3",
-    "dpdm": "3.10.0",
-    "istanbul-instrumenter-loader": "3.0.1",
-    "karma": "6.3.16",
+    "cross-var": "1.1.0",
+    "dpdm": "3.13.1",
+    "karma": "6.4.2",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-coverage": "2.2.1",
     "karma-mocha": "2.0.1",
     "karma-mocha-webworker": "1.3.0",
-    "karma-spec-reporter": "0.0.32",
+    "karma-spec-reporter": "0.0.36",
     "karma-webpack": "4.0.2",
-    "memfs": "3.4.9",
-    "mocha": "10.0.0",
+    "lerna": "6.6.2",
+    "memfs": "3.5.3",
+    "mocha": "10.2.0",
     "nyc": "15.1.0",
-    "sinon": "15.0.0",
+    "sinon": "15.1.2",
     "ts-loader": "8.4.0",
     "ts-mocha": "10.0.0",
     "typescript": "4.4.4",
-    "unionfs": "4.4.0",
-    "webpack": "4.46.0"
+    "unionfs": "4.5.1",
+    "webpack": "5.89.0"
   },
   "homepage": "https://github.com/open-telemetry/opentelemetry-js/tree/main/api",
   "sideEffects": false,
-  "gitHead": "56e6b1bb890f844b8963a146780d0b9cfa8abd0d"
+  "gitHead": "7be35c7845e206b27b682e8ce1cee850b09cec04"
 }
